Add files via upload
In the present research work, we aim to build efficient tools for predicting CKD occurrence, following an approach that exploits ML techniques. 

More specifically, first, we apply class balancing in order to tackle the non-uniform distribution of the instances in the two classes, then features ranking and analysis are performed, and finally, several ML models are trained and evaluated based on various performance metrics.